+++ service-managment/src/test/java/it/gov/pagopa/hubpa/servicemanagement/mock/TributeServiceModelMock.java
@@ -1,5 +1,6 @@
 package it.gov.pagopa.hubpa.servicemanagement.mock;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -17,25 +18,25 @@ public class TributeServiceModelMock {
 	tributeServiceModel.setIbanSecondary("IT76N0300203280879483594963");
 	tributeServiceModel.setIdPrimaryCreditor(1L);
 	tributeServiceModel.setIdSecondaryCreditor(2L);
-	tributeServiceModel.setPercentageSecondary(5);
+	tributeServiceModel.setPercentageSecondary(BigDecimal.valueOf(5));
 	List<InstallmentModel> listInstallment = new ArrayList<>();
 
 	InstallmentModel installmentModel = new InstallmentModel();
 	installmentModel.setDueDate(LocalDate.now().plusMonths(2));
-	installmentModel.setPercentagePrimary(20);
-	installmentModel.setPercentageSecondary(0);
+	installmentModel.setPercentagePrimary(BigDecimal.valueOf(20));
+	installmentModel.setPercentageSecondary(BigDecimal.valueOf(0));
 	listInstallment.add(installmentModel);
 
 	installmentModel = new InstallmentModel();
 	installmentModel.setDueDate(LocalDate.now().plusMonths(3));
-	installmentModel.setPercentagePrimary(50);
-	installmentModel.setPercentageSecondary(20);
+	installmentModel.setPercentagePrimary(BigDecimal.valueOf(50));
+	installmentModel.setPercentageSecondary(BigDecimal.valueOf(20));
 	listInstallment.add(installmentModel);
 
 	installmentModel = new InstallmentModel();
 	installmentModel.setDueDate(LocalDate.now().plusMonths(4));
-	installmentModel.setPercentagePrimary(30);
-	installmentModel.setPercentageSecondary(80);
+	installmentModel.setPercentagePrimary(BigDecimal.valueOf(30));
+	installmentModel.setPercentageSecondary(BigDecimal.valueOf(80));
 	listInstallment.add(installmentModel);
 
 	tributeServiceModel.setInstallments(listInstallment);
@@ -59,14 +60,14 @@ public class TributeServiceModelMock {
     public final static TributeServiceModel validationKOCase3() {
 	TributeServiceModel modelMock = TributeServiceModelMock.validationOKCase1();
 	modelMock.getInstallments().get(0).setDueDate(LocalDate.of(2021, 1, 1));
-	modelMock.getInstallments().get(0).setPercentageSecondary(5);
+	modelMock.getInstallments().get(0).setPercentageSecondary(BigDecimal.valueOf(5));
 	return modelMock;
     }
 
     public final static TributeServiceModel validationKOCase4() {
 	TributeServiceModel modelMock = TributeServiceModelMock.validationOKCase1();
-	modelMock.getInstallments().get(0).setPercentagePrimary(100);
-	modelMock.getInstallments().get(2).setPercentageSecondary(110);
+	modelMock.getInstallments().get(0).setPercentagePrimary(BigDecimal.valueOf(100));
+	modelMock.getInstallments().get(2).setPercentageSecondary(BigDecimal.valueOf(110));
 	return modelMock;
     }
 
@@ -79,13 +80,13 @@ public class TributeServiceModelMock {
 
     public final static TributeServiceModel validationKOCase6() {
 	TributeServiceModel modelMock = TributeServiceModelMock.validationOKCase1();
-	modelMock.setPercentageSecondary(0);
+	modelMock.setPercentageSecondary(BigDecimal.valueOf(0));
 	return modelMock;
     }
 
     public final static TributeServiceModel validationKOCase7() {
 	TributeServiceModel modelMock = TributeServiceModelMock.validationOKCase1();
-	modelMock.setPercentageSecondary(0);
+	modelMock.setPercentageSecondary(BigDecimal.valueOf(0));
 	modelMock.setDueDateUnique(LocalDate.of(2021, 1, 1));
 	modelMock.getInstallments().get(0).setDueDate(null);
 	modelMock.getInstallments().get(1).setPercentagePrimary(null);
@@ -104,10 +105,10 @@ public class TributeServiceModelMock {
     public final static TributeServiceModel validationOKCase2() {
 	TributeServiceModel modelMock = TributeServiceModelMock.validationOKCase1();
 	modelMock.setDueDateUnique(null);
-	modelMock.setPercentageSecondary(0);
-	modelMock.getInstallments().get(0).setPercentageSecondary(0);
-	modelMock.getInstallments().get(1).setPercentageSecondary(0);
-	modelMock.getInstallments().get(2).setPercentageSecondary(0);
+	modelMock.setPercentageSecondary(BigDecimal.valueOf(0));
+	modelMock.getInstallments().get(0).setPercentageSecondary(BigDecimal.valueOf(0));
+	modelMock.getInstallments().get(1).setPercentageSecondary(BigDecimal.valueOf(0));
+	modelMock.getInstallments().get(2).setPercentageSecondary(BigDecimal.valueOf(0));
 	return modelMock;
     }
 
@@ -115,7 +116,7 @@ public class TributeServiceModelMock {
 	TributeServiceModel modelMock = TributeServiceModelMock.validationOKCase1();
 	modelMock.setInstallments(null);
 	modelMock.setDueDateUnique(LocalDate.of(2021, 1, 1));
-	modelMock.setPercentageSecondary(0);
+	modelMock.setPercentageSecondary(BigDecimal.valueOf(0));
 	return modelMock;
     }
 }
